2.2.22 fix hover import and includes with missing extension
